Fix superanimal_humanbody unit test
